+++ frontend/src/App.tsx
@@ -16,6 +16,7 @@ import { SeoTrendChart } from './components/SeoTrendChart';
 import { SeoCompareView } from './components/SeoCompareView';
 import { TopicsPanel } from './components/TopicsPanel';
 import { CitationsPanel } from './components/CitationsPanel';
+import { ListicleGapPanel } from './components/ListicleGapPanel';
 import { CompetitorTrendChart } from './components/CompetitorTrend';
 import { AlertSettings } from './components/AlertSettings';
 import { SovChart } from './components/SovChart';
@@ -182,6 +183,11 @@ export default function App() {
                     <ResultTable results={scanResult.results} stats={scanResult.stats} />
                     <TopicsPanel results={scanResult.results} />
                     <CitationsPanel results={scanResult.results} brand={scanMeta.brand} />
+                    <ListicleGapPanel
+                      brand={scanMeta.brand}
+                      category={scanMeta.category}
+                      results={scanResult.results}
+                    />
                     <PromptCoverageMap brand={scanMeta.brand} />
                     <RecommendationsPanel recommendations={scanResult.recommendations} />
                     <ImpactPredictor results={scanResult.results} stats={scanResult.stats} brand={scanMeta.brand} />